+++ src/actions/index.ts
@@ -3,6 +3,8 @@ import { systemActions } from './system.js'
 import { autoFramingActions } from './autoframing.js'
 import { ptzActions } from './ptz.js'
 import { focusActions } from './focus.js'
+import { imagingActions } from './imaging.js'
+import { tallyActions } from './tally.js'
 import { presetActions } from './preset.js'
 import { sceneFileActions } from './scenefile.js'
 import { miscActions } from './misc.js'
@@ -16,6 +18,8 @@ export function UpdateActions(self: ModuleInstance): void {
 		...autoFramingActions(self),
 		...ptzActions(self),
 		...focusActions(self),
+		...imagingActions(self),
+		...tallyActions(self),
 		...presetActions(self),
 		...sceneFileActions(self),
 		...miscActions(self),